requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
Django==4.0.5
django-crontab==0.7.1
djangorestframework==3.13.1
fonttools==4.34.4
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
